requirements.txt
requests
beautifulsoup4
feedparser
pandas
gspread
google-auth
google-auth-oauthlib
google-api-python-client
APScheduler 